Split Util into Util and halp/Util and make package private
